+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+## Unreleased
+### Added
+- PageTemplate view to render a page template (including asset bundles, etc).
+
+## 1.0.0-rc.1 - 2017-11-26
+### Added
+- Requirement for Craft RC1 release
+
 ## 1.0.0-beta - 2017-03-22
 ### Added
 - Initial release.